refactor: extract SubList's rows, thumbnails and delete dialog (Q10)
Bulk thumbnail fetching and expiry scheduling move to
useThumbnailUrls, the /getsub window to useSubListRows, and the confirm
dialog becomes SubListDeleteDialog. Select-all is derived from the rows
on screen instead of mirrored by two effects, which also closes a
stale-true edge when a page change pruned the last checked key.
1,038 lines down to 769.
